### 2025-01-21 #7704 @editorialbot
[PRE REVIEW]: AquaCrop.jl: A Process-Based Model of Crop Growth
**Submitting author:** @gabo-di (<a href="http://orcid.org/0000-0001-6011-6097">Gabriel Diaz Iturry</a>)
**Repository:** https://github.com/gabo-di/AquaCrop.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.0
