### 2025-01-22 #7704 @veddox
**Reviewer suggestions:**

*From JOSS reviewer list:* Rémi Vezy, tonyewong, Julien Jean Malard-Adam

*Other potential reviewers:* Louise Busschaert (University of Leuven, codeveloper on the original Fortran version of the software), Han Su (University of Twente, has expressed interest in using AquaCrop.jl), Martijn Visser (Deltares, has expressed interest in using AquaCrop.jl)
