### 2025-01-28 #7704 @gabo-di
Reviewer suggestions:

- From JOSS:
  1. (@)VEZY
  2. (@)tonyewong
  3. (@)julienmalard

- Other potential reviewers:
  1. Louise Busschaert (@)lbusschaert (University of Leuven, codeveloper on the original Fortran version of the software).
  2. Martijn Visser (@)visr (Deltares, has expressed interest in using AquaCrop.jl)
  3. Han Su (email: h.su@utwente.nl)) (University of Twente, has expressed interest in using AquaCrop.jl)
