+++ README_EXTENDED_TESTS.md
@@ -1,0 +1,9 @@
+# Extended Tests
+Here we have some additional tests.
+We have additional data and comparisions to Fortran outputs.
+
+DO NOT MAKE A PR TO MAIN FROM THIS BRANCH!
+If some tests does not pass write the code in another branch like `dev-stuff` 
+then merge `dev-stuff` to this branch `extended-tests`, and make sure that all the tests pass.
+the make a PR from `dev-stuff`, make sure to do not pass the additinal data and tests to the 
+`main` branch.
