+++ test/extended_test/checkpoints.jl
@@ -1,0 +1,3176 @@
+using AquaCrop 
+
+function checkpoint1()
+    # Local variables
+
+    simulparam = AquaCrop.RepParam(
+        EvapDeclineFactor=4,
+        KcWetBare=1.1,
+        PercCCxHIfinal=5,
+        RootPercentZmin=70,
+        MaxRootZoneExpansion=5,
+        KsShapeFactorRoot=-6,
+        TAWGermination=20,
+        pAdjFAO=1,
+        DelayLowOxygen=3,
+        ExpFsen=1,
+        Beta=12,
+        ThicknessTopSWC=10,
+        EvapZmax=30,
+        RunoffDepth=0.3,
+        CNcorrection=true,
+        Tmin=12,
+        Tmax=28,
+        GDDMethod=3,
+        PercRAW=50,
+        CompDefThick=0.1,
+        CropDay1=81,
+        Tbase=10,
+        Tupper=30,
+        IrriFwInSeason=100,
+        IrriFwOffSeason=100,
+        ShowersInDecade=fill(AquaCrop.undef_int, 12),
+        EffectiveRain=AquaCrop.RepEffectiveRain(
+            method=:USDA, #1,
+            PercentEffRain=70,
+            ShowersInDecade=2,
+            RootNrEvap=5
+        ),
+        SaltDiff=20,
+        SaltSolub=100,
+        ConstGwt=true,
+        RootNrDF=16,
+        IniAbstract=5
+    )
+
+    soil = AquaCrop.RepSoil(
+        REW=9,
+        NrSoilLayers=1,
+        CNValue=61,
+        RootMax=1
+    )
+
+    soil_layers = AquaCrop.SoilLayerIndividual[
+        AquaCrop.SoilLayerIndividual(
+        Description="Loamy soil horizon",
+        Thickness=4,
+        SAT=50,
+        FC=30,
+        WP=10,
+        tau=0.76,
+        InfRate=500,
+        Penetrability=100,
+        GravelMass=0,
+        GravelVol=0,
+        WaterContent=360,
+        Macro=30,
+        SaltMobility=[0.030653430031715494, 0.99999999999999989, 1, 1, AquaCrop.undef_double, AquaCrop.undef_double, AquaCrop.undef_double, AquaCrop.undef_double, AquaCrop.undef_double, AquaCrop.undef_double, AquaCrop.undef_double], # [0.030653430031715494, 0.99999999999999989, 1, 1, 0, 0, 0, 0, 0, 0, 0],
+        SC=3,
+        SCP1=4,
+        UL=0.29999999999999999,
+        Dx=0.099999999999999992,
+        SoilClass=2,
+        CRa=-0.4536,
+        CRb=0.83734
+    )
+    ]
+
+    compartments = AquaCrop.CompartmentIndividual[
+        AquaCrop.CompartmentIndividual(
+            Thickness=0.10000000000000001,
+            Theta=0.29999999999999999,
+            Fluxout=AquaCrop.undef_double,#0,
+            Layer=1,
+            Smax=AquaCrop.undef_double,#0,
+            FCadj=30,
+            DayAnaero=0,
+            WFactor=AquaCrop.undef_double, #0,
+            Salt=zeros(Float64, 11),
+            Depo=zeros(Float64, 11),
+        )
+        for _ in 1:12
+    ]
+
+    simulation = AquaCrop.RepSim(
+        FromDayNr=81,
+        ToDayNr=205,
+        IniSWC=AquaCrop.RepIniSWC(
+            AtDepths=false,
+            NrLoc=1,
[…]
