+++ test/extended_test/checkpoints.jl
@@ -677,17 +677,9 @@ function checkpoint1()
     )
 end
 
-function checkpoint2()
+function checkpoint_project1_2()
     gvars = checkpoint1()    
 
-    gvars[:simulation].MultipleRun = true
-    gvars[:simulation].NrRuns = 3
-    gvars[:simulation].MultipleRunWithKeepSWC = true
-    gvars[:simulation].MultipleRunConstZrx = 3
-    # OJO this is incorrect in fortran code, they forget to set the temperature in line startuni.f90:864
-    # it should be: call SetSimulParam_Tmin(Tmin_temp)  corrected in v7.2
-    # gvars[:simulparam].Tmin = 0 
-    
     AquaCrop.setparameter!(gvars[:integer_parameters], :outputaggregate, 0)
 
     AquaCrop.setparameter!(gvars[:symbol_parameters], :theprojecttype, :typeprm)
@@ -721,65 +713,48 @@ function checkpoint2()
     AquaCrop.setparameter!(gvars[:string_parameters], :clim_file,  "EToRainTempFile")
     AquaCrop.setparameter!(gvars[:string_parameters], :swcini_file,  "(None)")
 
-    # fileok = AquaCrop.RepFileOK(
-    #     Climate_Filename=true,
-    #     Temperature_Filename=true,
-    #     ETo_Filename=true,
-    #     Rain_Filename=true,
-    #     CO2_Filename=true,
-    #     Calendar_Filename=true,
-    #     Crop_Filename=true,
-    #     Irrigation_Filename=true,
-    #     Management_Filename=true,
-    #     GroundWater_Filename=true,
-    #     Soil_Filename=true,
-    #     SWCIni_Filename=true,
-    #     OffSeason_Filename=true,
-    #     Observations_Filename=true
-    # )
-
     gvars[:projectinput] = AquaCrop.ProjectInputType[
         AquaCrop.ProjectInputType(
-            ParentDir=pwd()*"/testcase",
+            ParentDir=pwd()*"/extended_test/fortranrun",
             VersionNr=7.1,
-            Description="Ottawa (Canada)",
+            Description="UHK (Germany)",
             Simulation_YearSeason=1,
-            Simulation_DayNr1=41414,
-            Simulation_DayNrN=41577,
-            Crop_Day1=41414,
-            Crop_DayN=41577,
+            Simulation_DayNr1=41750,
+            Simulation_DayNrN=42299,
+            Crop_Day1=41757,
+            Crop_DayN=42158,
             Climate_Info="-- 1. Climate (CLI) file",
-            Climate_Filename="Ottawa.CLI",
+            Climate_Filename="uhk.CLI",
             Climate_Directory="DATA",
             Temperature_Info="1.1 Temperature (Tnx or TMP) file",
-            Temperature_Filename="Ottawa.Tnx",
+            Temperature_Filename="uhk.Tnx",
             Temperature_Directory="DATA",
             ETo_Info="1.2 Reference ET (ETo) file",
-            ETo_Filename="Ottawa.ETo",
+            ETo_Filename="uhk.ETo",
             ETo_Directory="DATA",
             Rain_Info="1.3 Rain (PLU) file",
-            Rain_Filename="Ottawa.PLU",
+            Rain_Filename="uhk.PLU",
             Rain_Directory="DATA",
             CO2_Info="1.4 Atmospheric CO2 concentration (CO2) file",
             CO2_Filename="MaunaLoa.CO2",
             CO2_Directory="SIMUL",
             Calendar_Info="-- 2. Calendar (CAL) file",
-            Calendar_Filename="21May.CAL",
-            Calendar_Directory="DATA",
+            Calendar_Filename="(None)",
[…]
