### 2025-02-04 bafc3ed
Update extended-tests.yml
error in last line
