+++ .github/workflows/extended-tests.yml
@@ -45,4 +45,3 @@ jobs:
       - uses: julia-actions/cache@v2
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
-          doctest(AquaCrop; manual=false)
