### 2025-02-05 9bada14
Add parsing function for daily output file of original Fortran code
