### 2025-02-06 dd8b782
Relax version testing to 7.x and allow for variable number of columns
