+++ src/readoutput.jl
@@ -9,32 +9,38 @@ Reads the daily output file produced by the original AquaCrop Fortran
 code (e.g. `OttawaPROday.OUT`) and returns the contents as a
 `DataFrame`.
 
-Note that this function cannot read the "PRM" dayily output file
-format that contains output for multiple runs in one file.
+Note: this function cannot read the "PRM"-style (`OttawaPRMday.OUT`)
+      daily output file format that contains output for multiple runs
+      in one file.
 """
 function read_orig_ac_output_dayout(io::IO)
-    # Note: There are two header lines.  The first line typically has
-    #       the name of the column, and the second line the unit.
-    #       Some columns such as "Day" or "Month" don't have a unit,
-    #       and some column names such as "WC 2" contain a space.
-    #       Luckily, all column names with a space are of the form
-    #       "<colname> <number>", so it is easy to fix.
+    # Note: There are four header lines, with the first two being just
+    #       a preamble.  The third line has the name of the column,
+    #       and the fourth line the column unit.  Some columns such as
+    #       "Day" or "Month" don't have a unit, and some column names
+    #       such as "WC 2" contain a space.  Luckily, all column names
+    #       with a space are of the form "<colname> <number>", so we
+    #       can fix it up after splitting the line on whitespace.
     #
     # TODO: Column names in the header "WC01" to "WC12" and "ECe01" to
     #       "ECe12" have a number below them, where other column names
     #       have the units (e.g. "ton/ha").  Currently, these numbers
     #       are simply added to the column names.
+
+    # Number of lines in the file header
     nheaderlines = 4
-    nfield = 98
-    nfield_hdr1 = 114  # before fixup
-    nfield_hdr2 = 93   # before fixup
+    # These columns are missing on the third header line, these are
+    # "Day", "Month", "Year", "DAP", "Stage" in the third header line
+    ncol_units_missing = 5
+    # Minimum number of columns we expect
+    ncol_min = ncol_units_missing
 
     lines = readlines(io)
     if length(lines) < nheaderlines
         throw(ArgumentError("File must contain at least $nheaderlines lines"))
     end
-    if lines[1][1:12] != "AquaCrop 7.2"
-        throw(ArgumentError("Daily output file must start with \"AquaCrop 7.2\" on the first line."))
+    if lines[1][1:11] != "AquaCrop 7."
+        throw(ArgumentError("Daily output file must start with \"AquaCrop 7.\" on the first line."))
     end
     if lines[2] != ""
         throw(ArgumentError("Expected empty second line."))
@@ -42,41 +48,41 @@ function read_orig_ac_output_dayout(io::IO)
     if occursin(r"^\s*Run:\s*\d+\s*", lines[3])
         throw(ArgumentError("This is a multi-run output file (PRM file, contains a \"Run:\" line), this function reads a single run (PRO file)"))
     end
-    # read header line 1
-    hdr1 = split(lines[3])
-    if length(hdr1) != nfield_hdr1
-        throw(ArgumentError("Unexpected number of words on first header line (line 3): expected $nfield_hdr1, got $(length(hdr1))."))
-    end
-    # For header line 1, some fields are unfortunately split by
-    # whitespace (e.g. "WC 1"), where the second part is always a
-    # number.  This loop joins these fields back together.
+
+    # Read third header line that contains the column names.  Column
+    # names are split on whitespace, with the additional twist that
+    # some columns names such as "WC 1" can also contain a space.  The
+    # column names containing spaces always have a number in the
[…]
