### 2025-02-10 3289d9c
Shortened manuscript as per editor's request
