+++ paper/paper.bib
@@ -284,24 +284,6 @@
         abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
 }
 
-@article{Danisch2021,
-  title = {Makie.Jl: {{Flexible}} High-Performance Data Visualization for {{Julia}}},
-  shorttitle = {Makie.Jl},
-  author = {Danisch, Simon and Krumbiegel, Julius},
-  date = {2021-09-01},
-  journaltitle = {Journal of Open Source Software},
-  volume = {6},
-  number = {65},
-  pages = {3349},
-  issn = {2475-9066},
-  doi = {10.21105/joss.03349},
-  url = {https://joss.theoj.org/papers/10.21105/joss.03349},
-  urldate = {2023-10-10},
-  abstract = {Danisch et al., (2021). Makie.jl: Flexible high-performance data visualization for Julia. Journal of Open Source Software, 6(65), 3349, https://doi.org/10.21105/joss.03349},
-  langid = {english},
-  file = {/home/xo30xoqa/Documents/Bibliothek/Zotero Library/Journal of Open Source Software/Danisch,Krumbiegel-2021-Makie.pdf}
-}
-
 @article{Foley2011,
   title = {Solutions for a Cultivated Planet},
   author = {Foley, Jonathan A. and Ramankutty, Navin and Brauman, Kate A. and Cassidy, Emily S. and Gerber, James S. and Johnston, Matt and Mueller, Nathaniel D. and O’Connell, Christine and Ray, Deepak K. and West, Paul C. and Balzer, Christian and Bennett, Elena M. and Carpenter, Stephen R. and Hill, Jason and Monfreda, Chad and Polasky, Stephen and Rockström, Johan and Sheehan, John and Siebert, Stefan and Tilman, David and Zaks, David P. M.},
