+++ paper/paper.md
@@ -50,37 +50,32 @@ interdisciplinary models in the environmental sciences.
 
 # Statement of need
 
-All agriculture is dependent on the growth of plants. Crop plants provide food
-for humans, fodder for domestic animals, and fibre and other resources for our
-manufacturing economy. Therefore, understanding how plants grow under different
-conditions is important not just for farmers, but also for the rest
-of society. This is particularly true in light of the major challenges associated with
-malnutrition and agriculture-related environmental degradation, both of which require
-wide-ranging changes to our food systems [@Foley2011]. In this context, crop models 
-that are based on physical and physiological processes can be used to inform 
-decision-making. These models use information about environmental parameters (e.g. 
-temperature, rainfall, soil quality) and knowledge of plant biology to simulate how 
-crop plants grow over time and estimate the resulting yield.
+All agriculture is dependent on the growth of plants, which provide food, fodder, 
+fibre, and other resources. Therefore, a detailed understanding of plant growth is
+vital for farmers, but also necessary to address the major food system challenges
+of our society. These include widespread malnutrition, agriculture-related
+environmental degradation, and global impacts of climate change, and require
+wide-ranging changes to our food systems [@Foley2011]. In this context, process-based 
+crop models can be used to inform decision-making. These models use information about 
+environmental parameters (e.g. temperature, rainfall, soil quality) and knowledge of 
+plant biology to simulate crop growth over time and estimate yield.
 
-As one such model, AquaCrop was developed with a special emphasis on the role
-of water for crop growth, and intended to be both simple and robust [@Steduto2009].
+AquaCrop is one such model. It lays a special emphasis on the role of water for crop 
+growth, and is intended to be both simple and robust [@Steduto2009].
 It has been used to model numerous crops worldwide [@Mialyk2024], and is known to
 produce reliable estimates of crop phenology and yield [@Kostkova2021]. First
 implemented in Delphi, it was later open-sourced in a Fortran version [@deRoos2021; 
 @RSDA2024]. There are also versions available in [Matlab](https://github.com/aquacropos/aquacrop-matlab),
 [Python](https://github.com/aquacropos/aquacrop), and [R](https://github.com/jrodriguez88/aquacrop-R),
-although currently these are not up-to-date with the most recent version of the original model 
-[@Foster2017; @Kelly2021; @CamargoRodriguez2019]. 
+although currently these are not up-to-date with the most recent version of the 
+original model [@Foster2017; @Kelly2021; @CamargoRodriguez2019]. 
 
 `AquaCrop.jl` expands this portfolio to contribute to the emerging ecosystem of
 environmental research software in Julia. To our knowledge, this is the first 
-process-based crop model available in this language. In addition, our purpose is also 
-to provide a package that can be readily integrated into other scientific software. 
-Recent research has emphasised the need for the creation of interdisciplinary models 
-that address the multifaceted processes underlying global challenges, such as climate 
-change or biodiversity loss [@Cabral2023]. This will require the use of model coupling
-and the adaptation of existing models to be usable as components in integrated models 
-[@Vedder2024].
+process-based crop model available in this language. Our purpose is also to provide a 
+package that can be readily integrated into other scientific software, in order to
+facilitate the creation of multidisciplinary models of socio-environmental systems
+[@Cabral2023,@Vedder2024].
 
 Specifically, we developed the package to use it as a component within 
 [`Persefone.jl`](https://persefone-model.eu), a process-based model of agricultural 
@@ -90,62 +85,39 @@ crop plants is an important mediating factor.
 
 # Comparison to original implementation
 
[…]
