### 2025-02-10 21152ad
Further shortening of the manuscript
