### 2025-02-10 #7704 @veddox
> Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly. Currently there are 1390 words.

Hello @kthyng, we've reduced the manuscript to 1067 words. Is that within the margin of error? As we are reporting the reimplementation of an existing model, we think it's important to show where we adapted its functionality, even though this makes for a longer section of text that doesn't usually appear in a JOSS paper.
