+++ test/extended_test/checkpoints.jl
@@ -1,6 +1,6 @@
 using AquaCrop 
 
-function checkpoint1()
+function checkpoint0()
     # Local variables
 
     simulparam = AquaCrop.RepParam(
@@ -677,12 +677,12 @@ function checkpoint1()
     )
 end
 
-function checkpoint_project1_2()
-    gvars = checkpoint1()    
+function checkpoint_project1_1()
+    gvars = checkpoint0()    
 
     AquaCrop.setparameter!(gvars[:integer_parameters], :outputaggregate, 0)
 
-    AquaCrop.setparameter!(gvars[:symbol_parameters], :theprojecttype, :typeprm)
+    AquaCrop.setparameter!(gvars[:symbol_parameters], :theprojecttype, :typepro)
 
     AquaCrop.setparameter!(gvars[:bool_parameters], :out1Wabal, true)
     AquaCrop.setparameter!(gvars[:bool_parameters], :out2Crop, true)
@@ -697,20 +697,20 @@ function checkpoint_project1_2()
     AquaCrop.setparameter!(gvars[:bool_parameters], :part2Eval, true)
 
     AquaCrop.setparameter!(gvars[:string_parameters], :man_file, 
-            joinpath(pwd(), "testcase/DATA/Ottawa.MAN"))
+            "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :observations_file, 
-            joinpath(pwd(), "testcase/OBS/Ottawa.OBS"))
+            "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :rain_file, 
-            joinpath(pwd(), "testcase/DATA/Ottawa.PLU"))
+            "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :temperature_file, 
-            joinpath(pwd(), "testcase/DATA/Ottawa.Tnx"))
+            "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :eto_file, 
-            joinpath(pwd(), "testcase/DATA/Ottawa.ETo"))
+            "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :prof_file, 
-            joinpath(pwd(), "testcase/DATA/Ottawa.SOL"))
+            "DEFAULT.SOL")
     AquaCrop.setparameter!(gvars[:string_parameters], :CO2_file, 
-            joinpath(pwd(), "testcase/SIMUL/MaunaLoa.CO2"))
-    AquaCrop.setparameter!(gvars[:string_parameters], :clim_file,  "EToRainTempFile")
+            "MaunaLoa.CO2")
+    AquaCrop.setparameter!(gvars[:string_parameters], :clim_file,  "(None)")
     AquaCrop.setparameter!(gvars[:string_parameters], :swcini_file,  "(None)")
 
     gvars[:projectinput] = AquaCrop.ProjectInputType[
@@ -771,8 +771,8 @@ function checkpoint_project1_2()
     return gvars 
 end
 
-function checkpoint_project1_3()
-    gvars = checkpoint_project1_2()
+function checkpoint_project1_2()
+    gvars = checkpoint_project1_1()
     outputs = AquaCrop.start_outputs()
 
     gvars[:soil].REW = 9
@@ -1262,6 +1262,17 @@ function checkpoint_project1_3()
     AquaCrop.setparameter!(gvars[:float_parameters], :previoussumgdd, 0) 
     AquaCrop.setparameter!(gvars[:float_parameters], :cdctotal, 0.01442417843464988) 
 
+    AquaCrop.setparameter!(gvars[:string_parameters], :rain_file, 
+            joinpath(pwd(), "extended_test/fortranrun/DATA/uhk.PLU"))
+    AquaCrop.setparameter!(gvars[:string_parameters], :temperature_file, 
[…]
