+++ src/loadvardict.jl
@@ -312,7 +312,7 @@ possible crop_type are
 ["maize", "wheat", "cotton", "alfalfaGDD", "barley", "barleyGDD", "cottonGDD", "drybean", "drybeanGDD",
 "maizeGDD", "wheatGDD", "sugarbeet", "sugarbeetGDD", "sunflower", "sunflowerGDD", "sugarcane",
 "tomato", "tomatoGDD", "potato", "potatoGDD", "quinoa", "tef", "soybean", "soybeanGDD",
-"sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat"]
+"sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat", "rapeseedGDD"]
 """
 function set_crop!(crop::RepCrop, crop_type::AbstractString; aux::Union{AbstractDict,Nothing}=nothing)
     if crop_type == "maize"
@@ -2780,6 +2780,92 @@ function set_crop!(crop::RepCrop, crop_type::AbstractString; aux::Union{Abstract
         crop.Assimilates.Period = 0         # Number of days at end of season during which assimilates are stored in root system
         crop.Assimilates.Stored = 0         # Percentage of assimilates transferred to root system at last day of season
         crop.Assimilates.Mobilized = 0         # Percentage of stored assimilates transferred to above ground parts in next season
+    elseif crop_type == "rapeseedGDD"
+        # based on default Sunflower acea_crops.py and 
+        # Zeleke, K. T., Luckett, D. & Cowley, R. Calibration and Testing of the FAO AquaCrop Model for Canola. Agronomy Journal 103, 1610–1618 (2011).
+        # for the GDDays data we used the data from CA4Py stations in rapeseed, and filled GDDCGC GDDCDC GDDaysToHIo with the values from wheatGDD
+        # skip this line 0         # File protected
+        crop.subkind = :Grain # 2         #.. fruit/grain producing crop
+        crop.Planting = :Seed # 1         #.. Crop is sown
+        crop.ModeCycle = :GDDays # 0         #.. Determination of crop cycle # by calendar days
+        crop.pMethod = :FAOCorrection # 1         #? Soil water depletion factors (p) are adjusted by ETo
+        crop.Tbase = 0.0       #.. Base temperature (°C) below which crop development does not progress
+        crop.Tupper = 30.0       #.. Upper temperature (°C) above which crop development no longer increases with an increase in temperature
+        # skip this line -9         # Total length of crop cycle in growing degree-days
+        crop.pLeafDefUL = 0.20      #.. Soil water depletion factor for canopy expansion (p-exp) - Upper threshold
+        crop.pLeafDefLL = 0.55      #.. Soil water depletion factor for canopy expansion (p-exp) - Lower threshold
+        crop.KsShapeFactorLeaf = 3.5       #.. Shape factor for water stress coefficient for canopy expansion (0.0 = straight line)
+        crop.pdef = 0.60      #.. Soil water depletion fraction for stomatal control (p - sto) - Upper threshold
+        crop.KsShapeFactorStomata = 5.0       #.. Shape factor for water stress coefficient for stomatal control (0.0 = straight line)
+        crop.pSenescence = 0.70      #.. Soil water depletion factor for canopy senescence (p - sen) - Upper threshold
+        crop.KsShapeFactorSenescence= 3.0       #.. Shape factor for water stress coefficient for canopy senescence (0.0 = straight line)
+        crop.SumEToDelaySenescence = 50         #? Sum(ETo) during dormant period to be exceeded before crop is permanently wilted
+        crop.pPollination = 0.85      #? Soil water depletion factor for pollination (p - pol) - Upper threshold
[…]
