+++ src/api.jl
@@ -422,7 +422,7 @@ The `crop_type` must be one of these  strings indicating the crop type:
 `["maize", "wheat", "cotton", "alfalfaGDD", "barley", "barleyGDD", "cottonGDD", "drybean", "drybeanGDD",
 "maizeGDD", "wheatGDD", "sugarbeet", "sugarbeetGDD", "sunflower", "sunflowerGDD", "sugarcane",
 "tomato", "tomatoGDD", "potato", "potatoGDD", "quinoa", "tef", "soybean", "soybeanGDD",
-"sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat"]`
+"sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat", "rapeseedGDD"]`
 
 We also have the optional keys:
 `[:co2i, :crop, :perennial_period, :soil, :soil_layers, :simulparam,
@@ -488,7 +488,7 @@ function check_nofilerun(outputs; kwargs...)
     crop_types = ["maize", "wheat", "cotton", "alfalfaGDD", "barley", "barleyGDD", "cottonGDD", "drybean", "drybeanGDD",
                     "maizeGDD", "wheatGDD", "sugarbeet", "sugarbeetGDD", "sunflower", "sunflowerGDD", "sugarcane",
                     "tomato", "tomatoGDD", "potato", "potatoGDD", "quinoa", "tef", "soybean", "soybeanGDD",
-                    "sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat"]
+                    "sorghum", "sorghumGDD", "paddyrice", "paddyriceGDD", "rapeseed", "oat", "rapeseedGDD"]
     if !haskey(kwargs, :crop_type) 
         all_ok.logi = false
         all_ok.msg = "missing  necessary keyword :crop_type"
