+++ src/initialsettings.jl
@@ -20,13 +20,13 @@ function check_for_keep_swc(outputs, projectinput::Vector{ProjectInputType}, fil
     filename = projectinput[1].Soil_Filename
     has_external = filename == "(External)"
 
-    if (has_external) 
+    if (has_external)
         # Note: here we use the AquaCrop version number and assume that
         # the same version can be used in finalizing the soil settings.
         soil = gvars[:soil]
         soil_layers = gvars[:soil_layers]
         compartments = gvars[:compartments]
-    elseif (filename == "(None)") 
+    elseif (filename == "(None)")
         soil = gvars[:soil]
         soil_layers = gvars[:soil_layers]
         compartments = gvars[:compartments]
@@ -34,21 +34,21 @@ function check_for_keep_swc(outputs, projectinput::Vector{ProjectInputType}, fil
         soil, soil_layers, compartments = load_profile(outputs,
             joinpath([filepaths[:prog], projectinput[1].Soil_Directory, filename]),
             gvars[:simulparam]; kwargs...)
-    end 
+    end
 
     # 3. Check if runs with KeepSWC exist
     runi = 1
     totalnrofruns = length(projectinput)
     while ((runwithkeepswc == false) & (runi <= totalnrofruns))
-        if (projectinput[runi].SWCIni_Filename == "KeepSWC") 
+        if (projectinput[runi].SWCIni_Filename == "KeepSWC")
             runwithkeepswc = true
-        end 
+        end
         runi = runi + 1
-    end 
+    end
 
-    if (runwithkeepswc == false) 
+    if (runwithkeepswc == false)
         constzrxforrun = undef_double # reset
-    end 
+    end
 
     # 4. Look for maximum root zone depth IF RunWithKeepSWC
     if (runwithkeepswc == true)
@@ -56,16 +56,16 @@ function check_for_keep_swc(outputs, projectinput::Vector{ProjectInputType}, fil
         while (runi <= totalnrofruns)
             # Obtain maximum rooting depth from the crop file
             fullfilename = joinpath([filepaths[:prog], projectinput[runi].Crop_Directory, projectinput[runi].Crop_Filename])
-        
+
             constzrxforrun = _check_for_keep_swc(kwargs[:runtype], fullfilename, soil_layers, constzrxforrun)
             runi = runi + 1
-        end 
-    end 
+        end
+    end
 
     # 5. Reload existing soil file is not necessary since we do not change any variables
 
     return runwithkeepswc, constzrxforrun
-end 
+end
 
 function check_for_keep_swc(outputs, projectinput::Vector{AbstractParametersContainer}, filepaths, gvars; kwargs...)
     return check_for_keep_swc(outputs, ProjectInputType[p for p in projectinput], filepaths, gvars; kwargs...)
@@ -77,14 +77,14 @@ function _check_for_keep_swc(runtype::NormalFileRun, fullfilename, soil_layers, 
         readline(file) # description
         versionnr = parse(Float64,strip(readline(file))[1:4])
         for i in 1:34
-            readline(file) 
-        end 
+            readline(file)
+        end
         zrni = parse(Float64,strip(readline(file))[1:6])
         zrxi = parse(Float64,strip(readline(file))[1:6])
         zrsoili = root_max_in_soil_profile(zrxi, soil_layers)
-        if (zrsoili > constzrxforrun) 
+        if (zrsoili > constzrxforrun)
             ret[1] = zrsoili
-        end 
+        end
     end
     return ret[1]
 end
[…]
