### 2025-03-24 #7704 @ethanwhite
@kthyng - I'd be happy to be the editor on this paper
