### 2025-03-24 #7704 @ethanwhite
Hi @veddox - I've done an initial check of the paper/software and everything looks good so I'm going to start looking for reviewers. This is my first time editing for JOSS, so apologies in advance for any hiccups along the way as I figure things out.
