### 2025-03-24 #7944 @editorialbot
[REVIEW]: AquaCrop.jl: A Process-Based Model of Crop Growth
**Submitting author:** @gabo-di (<a href="http://orcid.org/0000-0001-6011-6097">Gabriel Diaz Iturry</a>)
**Repository:** https://github.com/gabo-di/AquaCrop.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
