### 2025-04-04 ebfb62b
add function to save a tunned crop
