### 2025-04-07 #7944 @ethanwhite
:wave: @julienmalard & @tonyewong. Thanks again for agreeing to review this paper. I can see that @julienmalard's checklist is already making progress. Much appreciated! Since we're now two weeks into the review process, I just wanted to check in and see if either of you have any questions or have run into any issues that I can help out with as you work on your reviews.
