### 2025-04-12 #7944 @tonyewong
@veddox I'm reading over the paper. It looks great! I have just a couple things that came to mind as possible additions. I think 1 and 2 would be easy and useful to add, so I'll put issues in the repo for them. My #3 I'm fine either way you decide makes the most sense.

1. Line 31 - is the current model version/the one that's maintained regularly in Fortran? I suggest specifying what the current version is.
2. There's discussion of other AquaCrop versions, past and present, but are there other models or model families that are comparable? Would be useful to compare
3. I don't know that all of the code associated with the plotting for the example at line 63 is really needed. I'd be fine just seeing the output plot and a note in text about how these figures are a plot of direct model output. On the other hand, seeing the code communicates well that the model output is directly usable in that way. So, I'm happy seeing it either way, just my two cents!
