### 2025-04-14 b09cb3b
Added paragraph on other crop models
response to #42
