### 2025-04-14 #7944 @veddox
Hi @tonyewong, thanks for the good comments! I've addressed 1 & 2 in this commit: https://github.com/gabo-di/AquaCrop.jl/commit/b09cb3bd10c797e4fcbaf68cf2793f93671a8fd4

Regarding no. 3, my tendency would be to leave the code in. When getting to know a new package, I find it helpful to see an example of how to use it straight away. However, if @julienmalard also thinks we should remove it, we can of course do so.
