+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"CurrentModule = AquaCrop\nDocTestSetup = quote\n    using AquaCrop\nend","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"Pages = [\"gettingstarted.md\"]","category":"page"},{"location":"gettingstarted/#Install","page":"Getting Started","title":"Install","text":"","category":"section"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"To install the package, open Julia and run:","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"using Pkg\nPkg.add(url=\"https://github.com/gabo-di/AquaCrop.jl\")","category":"page"},{"location":"gettingstarted/#basic_run_section","page":"Getting Started","title":"Basic Run","text":"","category":"section"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"To maintain compatibility, we support the same input file formats used by the  original implementation of AquaCrop (v7.2). To specify that this format should be used, call the NormalFileRun function:","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"using AquaCrop\nruntype = NormalFileRun()","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"Next we must specify where these input files are found. In this case, we use the files in the test directory:","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"parentdir = AquaCrop.test_dir #\".../AquaCrop.jl/test/testcase\"\nendswith(parentdir, \"test/testcase\")","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"Now we are ready to use the basic_run function:","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"outputs = basic_run(; runtype=runtype, parentdir=parentdir);\n\n\nisequal(size(outputs[:dayout]), (892, 89))\n\n# output\ntrue","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"All model results are stored in a collection of dataframes. For example, we can  see the harvest dataframe via:","category":"page"},{"location":"gettingstarted/","page":"Getting Started","title":"Getting Started","text":"outputs[:harvestsout]\n\n# output\n14×11 DataFrame\n Row │ RunNr  Nr     Date        DAP    Interval   Biomass           Sum(B)    ⋯\n     │ Int64  Int64  Date        Int64  Quantity…  Quantity…         Quantity… ⋯\n─────┼──────────────────────────────────────────────────────────────────────────\n   1 │     1      0  2014-05-21      0      0.0 d      0.0 ton ha⁻¹      0.0 t ⋯\n   2 │     1      1  2014-07-13     54     54.0 d  4.78668 ton ha⁻¹  4.78668 t\n   3 │     1      2  2014-08-31    103     49.0 d  3.08189 ton ha⁻¹  7.86857 t\n   4 │     1   9999  2014-11-01      0      0.0 d      0.0 ton ha⁻¹  9.17245 t\n   5 │     2      0  2015-05-01      0      0.0 d      0.0 ton ha⁻¹      0.0 t ⋯\n   6 │    
[…]
