### 2025-04-16 #7944 @gabo-di
Hello @tonyewong  thank you for the comments.
1. The tests were done running [FAO's AquaCrop](https://github.com/KUL-RSDA/AquaCrop/) in debug mode with [GDB](https://en.wikipedia.org/wiki/GNU_Debugger). Every now and then we set breaking points and compared Fortran's output with  [AquaCrop.jl](https://github.com/gabo-di/AquaCrop.jl)'s output, (Fortran's outputs are recorded in [AquaCrop.jl/test/checkpoints.jl](https://github.com/gabo-di/AquaCrop.jl/blob/main/test/checkpoints.jl) and Julia's runs are done on the other test scripts). Maybe the whole test process is too convoluted to point out on the README.md. On the other hand, if those test do not go well it will be shown on the README.md in either the Continuous Integration Badge:  [![Build Status](https://github.com/gabo-di/AquaCrop.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/gabo-di/AquaCrop.jl/actions/workflows/CI.yml?query=branch%3Amain) or the Extended Tests Badge : [![Extended Tests](https://github.com/gabo-di/AquaCrop.jl/actions/workflows/extended-tests.yml/badge.svg)](https://github.com/gabo-di/AquaCrop.jl/actions/workflows/extended-tests.yml?query=branch%3Amain).
2. Thank you for pointing out this issue. We have updated the documentation and the readme with the following phrase "[...]  making it more easily accesible to growing number of interdisciplinary environmental modellers working with Julia." (@veddox  do you have anything to add in this point?)
Let me know if we addressed your comments correctly.
