### 2025-04-18 8b95741
Update README.md describing how to test package
