### 2025-04-18 7a2e420
Update README.md describing how to test package
