### 2025-04-18 #7944 @gabo-di
Hello @tonyewong  . We have added a short paragraph on the readme about how to run AquaCrop's tests
