### 2025-04-18 #7944 @ethanwhite
👋 @gabo-di @veddox @tonyewong @julienmalard. Thanks to all of you for your work so far on this review and I'd particularly like to commend you all on an excellent collaborative set of interactions. This is exactly how peer review should work.

It looks like responses to @tonyewong's review are almost complete and I agree with @veddox that keeping the code chunk in is valuable.

@julienmalard - Thanks for your work so far! We're now 4 weeks into the review and we like to get them done within 6 weeks if at all possible, so this is a friendly reminder that we've got 2 weeks left.
