### 2025-05-10 #7944 @julienmalard
@veddox @ethanwhite Apologies (again) for the delay; would this Tuesday be fine?
