### 2025-05-12 gabo-di/AquaCrop.jl#46 @julienmalard
JOSS Paper suggestions
Comments for the JOSS paper
* In the summary, consider mentionning what language AquaCrop was originally developped in.
* The section "comparison to original implementation" is very well-written.
* Perhaps justify the decision to translate the entire model to Julia, rather than implementing a wrapper-based approach that would interface with the original model and thus be easier to keep up to date.
* Regarding the discussion of figure 2 and that it shows that "when well parameterised, AquaCrop.jl forecasts the development of yields over time quite reliably," I would recommend including more information. How was the model parametrised? On this data set (in which case the results are not that remarkable), or on another dataset? Could you give a few examples of model fit statistics (e.g., R2, RMSE, KGE or other)? And most importantly, how do these results compare with the original AquaCrop implementation (ideally close to identical)?
* Consider publishing AquaCrop.jl to the Julia package registry.
