### 2025-05-12 #7944 @julienmalard
Hello,
I have published my reviewer comments in an issue on the AquaCrop.jl repository: https://github.com/gabo-di/AquaCrop.jl/issues/46
