### 2025-05-12 #7944 @ethanwhite
Thanks @julienmalard - much appreciated!

@veddox - I look forward to the resulting changes and getting this wrapped up soon!
