### 2025-05-12 00e2ece
Addressed reviewer comments (issue #46)
