### 2025-05-14 gabo-di/AquaCrop.jl#46 @gabo-di
Hello @julienmalard , thank you for your comment.
- We are mentioning that FAO's AquaCrop current version is in Fortran language, (mentioned both on the paper's summary as in the code's documentation).

- Thank you for that comment.

- The decision to translate the model was two main reasons:
1.  AquaCrop.jl is part of a bigger project for ecological simulations: [Persefone.jl](https://persefone-model.eu). We need a fine grained control of AquaCrop's simulations, where we can quickly and dynamically interact with the model's variables. FAO's AquaCrop can run day by day using the GUI, but not directly the github's Fortran repository. So a simple wrapper would not give us the amount of control on the simulation that we want (and have now), best case scenario a wrapper would run the whole process and then parse the Fortran output.
2. In the future, AquaCrop.jl can also be better integrated with Julia's growing ecosystem of composable numerical libraries. One possible example of this would be the use of automatic differentiation with AquaCrop.jl for gradient-based optimisation of crop parameters.

- We would like to avoid a long discussion of model fit, as we see this as beyond the scope of JOSS's paper, so we added a url reference to [CropGrowthTutorial](https://github.com/gabo-di/CropGrowthTutorial), where we discuss more about the parameter's calibration. Nevertheless, on the current paper we added a show case of how to use AquaCrop_jll, a high level wrapper that runs directly FAO's AquaCrop, to facilitate comparisons between both implementations.

- Thank you for the suggestion. We are addressing that here #47

@veddox  @marcom I took your ideas/comments to respond the issue.
