### 2025-05-14 #7944 @gabo-di
Thank you @julienmalard, we have answered your concerns on the same issue gabo-di/AquaCrop.jl#46
