### 2025-05-15 gabo-di/AquaCrop.jl#46 @julienmalard
Thanks; it looks very good to me!
