### 2025-05-16 #7944 @ethanwhite
Thanks for your excellent reviews @julienmalard and @tonyewong! And thanks to @veddox for promptly integrated feedback from both reviewers!

I've now read through the paper again and have just a couple of minor wording suggestions to consider:

- Line 28: "other" -> "others" since it is referring to more than one model
- Line 33: ", and shown" -> "and has been shown" or "and shown" (just dropping the comma)
- Line 44-45: I would combine these two paragraphs (the last two paragraphs of the Statement of Need) since they are both part of the same core idea. Alternatively remove the "Specifically," at the start of the second paragraph.

@veddox - once you've integrated these suggestions, then make a new release of the software, create a new archive of that release on Zenodo/figshare, and provide a link to the release on GitHub and the associated archive. Some of these steps will be listed on the post review checklist that I'll create in a minute.
