### 2025-05-25 0f17f2b
update version and installation instructions
