### 2025-05-27 #7944 @ethanwhite
Thanks @gabo-di!

The title and license of the archive should all match those associated with the paper and typically the author line matches as well. Please update the title and the license to match and either update the author list of the archive or explain why the author list of the paper should differ from the archive. Typically you can update the metadata of the archive on Zenodo without a new release, which you can do by logging into Zenodo and manually updating the relevant metadata fields for your current release. I'm not sure if you can also make changes to the license without a new release. If you can't, then a patch level release to facilitate this change is fine and you can move to `v1.0.1`. You can do this through GitHub by adding a [.zenodo.json file](https://help.zenodo.org/docs/github/describe-software/zenodo-json/) to your repo with the relevant title, authors, and license. Doing so will also make sure that changes to the title, authors, license, etc. persist with new releases.

I'm happy to help if you run into any issues while making these updates to the archive
