+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0000-0001-6011-6097
   - name: Marco C. Matthies
     affiliation: "1,2"
-	orcid: 0009-0006-9456-6786
+    orcid: 0009-0006-9456-6786
   - name: Guy Pe'er
     affiliation: "1,2"
     orcid: 0000-0002-7090-0560
