### 2025-05-29 #7944 @gabo-di
Hello @ethanwhite .

Thank you for your observations. We have updated the authors list, the license and the title on the zenodo release.
