### 2025-06-02 #7944 @gabo-di
Hello @ethanwhite , thank you for the observation. I have changed the author order on Zenodo.
