### 2025-06-02 #7944 @ethanwhite
@gabo-di - it looks like the Pan 2021 reference is included twice in `paper.bib`. Please remove one of them and I'll try the dry run again.
