+++ paper/paper.bib
@@ -302,25 +302,6 @@
   file = {/home/xo30xoqa/Documents/Bibliothek/Zotero Library/Nature/Foley et al-2011-Solutions for a cultivated planet.pdf}
 }
 	
-@incollection{Pan2021,
-  title = {Crop {{Growth Modeling}} and {{Yield Forecasting}}},
-  booktitle = {Agro-Geoinformatics: {{Theory}} and {{Practice}}},
-  author = {Pan, Haizhu and Chen, Zhongxin},
-  editor = {Di, Liping and {\"U}st{\"u}nda{\u g}, Berk},
-  year = {2021},
-  series = {Springer {{Remote Sensing}}/{{Photogrammetry}}},
-  pages = {205--220},
-  publisher = {Springer International Publishing},
-  address = {Cham},
-  doi = {10.1007/978-3-030-66387-2_11},
-  urldate = {2022-02-17},
-  abstract = {In the past decades, crop growth modeling and yield forecasting have attracted increasing attention in both scientific researches and agricultural practices. Many scientific studies have been carried out to improve the capabilities of crop growth modeling and yield forecasting by using various data sources and methods like statistical models, crop growth simulation models, and remote sensing. In this chapter, four categories of crop growth models were reviewed. Firstly, the traditional crop modeling and forecasting methods were introduced: statistical modeling and crop growth models. Then remote sensing models mainly based on spectral indices and quantitative products were introduced. The quality of remote sensing data is critical for crop modeling and yield forecasting. Finally, the widely used data assimilation of crops was described. More research is necessary for the full use of the value of remote sensing and crop growth model in crop growth monitoring and yield forecasting at a regional scale.},
-  isbn = {978-3-030-66387-2},
-  langid = {english},
-  keywords = {Crop growth,Data assimilation,Modeling,Remote sensing,Yield forecast},
-  file = {/home/xo30xoqa/Documents/Bibliothek/Zotero Library/Springer International Publishing/Pan,Chen-2021-Crop Growth Modeling and Yield Forecasting.pdf}
-}
-	
 @article{Holzworth2014,
   title = {{{APSIM}} -- {{Evolution}} towards a New Generation of Agricultural Systems Simulation},
   author = {Holzworth, Dean P. and Huth, Neil I. and {deVoil}, Peter G. and Zurcher, Eric J. and Herrmann, Neville I. and McLean, Greg and Chenu, Karine and {van Oosterom}, Erik J. and Snow, Val and Murphy, Chris and Moore, Andrew D. and Brown, Hamish and Whish, Jeremy P. M. and Verrall, Shaun and Fainges, Justin and Bell, Lindsay W. and Peake, Allan S. and Poulton, Perry L. and Hochman, Zvi and Thorburn, Peter J. and Gaydon, Donald S. and Dalgliesh, Neal P. and Rodriguez, Daniel and Cox, Howard and Chapman, Scott and Doherty, Alastair and Teixeira, Edmar and Sharp, Joanna and Cichota, Rogerio and Vogeler, Iris and Li, Frank Y. and Wang, Enli and Hammer, Graeme L. and Robertson, Michael J. and Dimes, John P. and Whitbread, Anthony M. and Hunt, James and {van Rees}, Harm and McClelland, Tim and Carberry, Peter S. and Hargreaves, John N. G. and MacLeod, Neil and McDonald, Cam and Harsdorf, Justin and Wedgwood, Sara and Keating, Brian A.},
